models/dtos are now seperated by feature to handle scaling.
