+++ Controllers/TagController.cs
@@ -1,5 +1,5 @@
 ﻿using Knowledge_Center_API.DataAccess.Demo;
-using Knowledge_Center_API.Models;
+using Knowledge_Center_API.Models.Tags;
 using Knowledge_Center_API.Services.Core;
 using Knowledge_Center_API.Services.Security;
 using Microsoft.AspNetCore.Mvc;
