+++ Models/KnowledgeNodes/KnowledgeNode.cs
@@ -1,13 +1,17 @@
-﻿namespace Knowledge_Center_API.Models.DTOs
+﻿namespace Knowledge_Center_API.Models.KnowledgeNodes
 {
-    public class KnowledgeNodeInlineDto
+    public class KnowledgeNode
     {
         public int Id { get; set; }
         public string Title { get; set; }
-        public string NodeType { get; set; }
+        public int DomainId { get; set; }
+        public string NodeType { get; set; } 
+        public string Description { get; set; }
         public int ConfidenceLevel { get; set; }
         public string Status { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime LastUpdated { get; set; }
+
+
     }
 }
