+++ Models/KnowledgeNodes/KnowledgeNodeWithLogsDto.cs
@@ -1,6 +1,15 @@
-﻿namespace Knowledge_Center_API.Models
+﻿using Knowledge_Center_API.Models.LogEntries;
+
+namespace Knowledge_Center_API.Models.KnowledgeNodes
 {
-    public class KnowledgeNode
+    /* 
+     * GET /api/knowledge-nodes/{id}
+     * 
+     * A KN has many logs, that is my design of this entity. 
+     * So to translate that, I added the LogEntryInline Dto 
+     * To this new KN Dto. 
+     */
+    public class KnowledgeNodeWithLogsDto
     {
         public int Id { get; set; }
         public string Title { get; set; }
@@ -12,6 +21,7 @@
         public DateTime CreatedAt { get; set; }
         public DateTime LastUpdated { get; set; }
 
+        public List<LogEntryInlineDto> Logs { get; set; }
 
     }
 }
