+++ Services/Core/DomainService.cs
@@ -1,6 +1,6 @@
 ﻿using Knowledge_Center_API.DataAccess;
-using Knowledge_Center_API.Models;
-using Knowledge_Center_API.Models.DTOs;
+using Knowledge_Center_API.Models.Domains;
+using Knowledge_Center_API.Models.KnowledgeNodes;
 using Knowledge_Center_API.Services.Validation;
 using Microsoft.Data.SqlClient;
 using System;
