+++ Services/Core/LogEntryService.cs
@@ -1,9 +1,9 @@
-﻿using Knowledge_Center_API.Models;
-using Knowledge_Center_API.Models.DTOs;
-using Knowledge_Center_API.Services.Validation;
+﻿using Knowledge_Center_API.Services.Validation;
 using Knowledge_Center_API.DataAccess;
 using Microsoft.Data.SqlClient;
 using System.Data;
+using Knowledge_Center_API.Models.LogEntries;
+using Knowledge_Center_API.Models.Tags;
 
 namespace Knowledge_Center_API.Services.Core
 {
