+++ Services/Core/UserService.cs
@@ -1,5 +1,5 @@
 ﻿using Knowledge_Center_API.DataAccess;
-using Knowledge_Center_API.Models;
+using Knowledge_Center_API.Models.Auth;
 using Knowledge_Center_API.Services.Security;
 using Knowledge_Center_API.Services.Validation;
 using Microsoft.Data.SqlClient;
